Add edit work and project to about tab in workplan page
