+++ epictrack-web/package.json
@@ -79,6 +79,7 @@
     "eslint": "^8.39.0",
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-prettier": "^4.2.1",
-    "prettier": "^2.8.8"
+    "prettier": "^2.8.8",
+    "prettier-eslint": "^16.3.0"
   }
 }
