+++ epictrack-web/src/components/workPlan/about/aboutDetails/ProjectDetails/Skeleton.tsx
@@ -1,0 +1,7 @@
+import { Skeleton } from "@mui/material";
+
+const ProjectDetailsSkeleton = () => {
+  return <Skeleton variant="rectangular" width="100%" height="100%" />;
+};
+
+export default ProjectDetailsSkeleton;
