+++ epictrack-web/src/components/workPlan/about/aboutDetails/WorkDetails/Skeleton.tsx
@@ -1,0 +1,7 @@
+import { Skeleton } from "@mui/material";
+
+const WorkDetailsSkeleton = () => {
+  return <Skeleton variant="rectangular" width="100%" height="100%" />;
+};
+
+export default WorkDetailsSkeleton;
